Added interfaces to apptypings
