+++ application-typings/src/Gemstone.ts
@@ -20,6 +20,7 @@
 //       Generated original version of source code.
 //
 // ******************************************************************************************************
+import * as React from 'react';
 
 namespace Gemstone {
     export namespace TSX {
@@ -31,6 +32,77 @@ namespace Gemstone {
                 Width: number,
                 Height: number
             }
+            export interface ICSVField<T> {
+                /**
+                 * The field in the record this definition applies to.
+                 * @type {keyof T}
+                 */
+                Field: keyof T;
+            
+                /**
+                 * The label for the field, used for select element.
+                 * @type {string}
+                 */
+                Label: string;
+            
+                /**
+                 * Function to validate the field value.
+                 * @param {string} value - The value to validate.
+                 * @returns {boolean}
+                 */
+                Validate: (value: string) => boolean;
+            
+                /**
+                 * Component for editing the field value.
+                 */
+                EditComponent: React.FC<{ Value: string, SetValue: (val: string) => void, Valid: boolean, Feedback?: string }>;
+            
+                /**
+                 * Optional help text for the select element.
+                 * @type {string}
+                 * @optional
+                 */
+                Help?: string;
+            
+                /**
+                 * Optional feedback for the EditComponent
+                 * @type {string}
+                 * @optional
+                 */
+                Feedback?: string;
+            
+                /**
+                 * Function to process the field value and update the record.
+                 * @param {string} val - The value to process.
+                 * @param {T} record - The record to update.
+                 * @returns {T}
+                 */
+                Process: (val: string, record: T) => T;
+            
+                /**
+                 * Flag indicating if the field is required.
+                 * @type {boolean}
+                 */
+                Required: boolean;
+            
+                /**
+                 * Flag indicating if the field can be empty.
+                 * @type {boolean}
+                 */
+                AllowEmpty: boolean;
+            
+                /**
+                 * Flag indicating if the field values must be unique.
+                 * @type {boolean}
+                 */
+                Unique: boolean;
+            }
+            export interface ISearchFilter<T> {
+                FieldName: keyof T,
+                SearchParameter: string,
+                Operator: ('=' | '<>' | '>' | '<' | '>=' | '<=' | 'LIKE' | 'NOT LIKE' | 'IN' | 'NOT IN')
+            }
+            
         }
     }
 }
