+++ react-interactive/src/index.tsx
@@ -47,7 +47,6 @@ import ConfigurableColumn from './ConfigurableTable/ConfigurableColumn';
 import GenericController from './GenericController';
 import LayoutGrid from './LayoutGrid';
 import UploadCSV from './UploadCSV';
-import { ICSVField } from './UploadCSV';
 
 const ConfigTable = {
   Table: ConfigurableTable,
@@ -79,6 +78,5 @@ export {
   BtnDropdown,
   GenericController,
   LayoutGrid,
-  UploadCSV,
-  ICSVField
+  UploadCSV
 };
